requested changes from code review (#73)
* update commit id of flir

* remove docs

* remove script folder

* rename TileIR to CommonIR

* revert unrelevant changes

* revert changes unrelated to ascend

* revert changes to backend testing tools

* revert unrelated create_tanh ffi

* revert compiler flags

* remove HIVM dependency of tle

* remove redundant header include directory

* guard commonir c++ changes with __TLE_DSA__

* add commonir python comments and guard

* change the license header for newly added source

* run commonir success tests in ascend ci

* move common ir unit test to it's own folder

* add more python comment

* remove envvar to replace ir

* move common ir changes to tle/dsa

* add more python comments

* add comment to explain flir commit id changes

---------
