+++ .github/workflows/ascend3.5-910b-tle-test.yml
@@ -104,3 +104,6 @@ jobs:
           python3 07-kv-rmsnorm-rope-cache.py
           python3 -m pytest -k "accuracy" 08-add-rms-norm-bias.py
           popd
+          pushd test/CommonIR/
+          bash Ascend/success_case/run_all_tests.sh
+          popd
