+++ bin/CMakeLists.txt
@@ -1,10 +1,5 @@
 get_property(triton_libs GLOBAL PROPERTY TRITON_LIBS)
 
-include_directories(
-  ${CMAKE_SOURCE_DIR}/third_party/flir/include
-  ${CMAKE_BINARY_DIR}/third_party/flir/include
-)
-
 add_llvm_executable(triton-opt triton-opt.cpp PARTIAL_SOURCES_INTENDED)
 
 # TODO: what's this?
