+++ bin/RegisterTritonDialects.h
@@ -18,8 +18,6 @@
 #include "third_party/tle/dialect/include/Transforms/Passes.h"
 #include "tle/dialect/include/IR/Dialect.h" // flagtree tle raw
 #endif
-#include "bishengir/Dialect/HIVM/IR/HIVM.h"
-#include "mlir-ext/Dialect/TileIR/IR/TileIRDialect.h"
 #include "triton/Dialect/Gluon/Transforms/Passes.h"
 #include "triton/Dialect/Triton/IR/Dialect.h"
 #include "triton/Dialect/TritonGPU/IR/Dialect.h"
@@ -138,7 +136,6 @@ inline void registerTritonDialects(mlir::DialectRegistry &registry) {
 
   registry.insert<
       mlir::triton::TritonDialect, mlir::cf::ControlFlowDialect,
-      mlir::triton::tile::TileIRDialect, mlir::hivm::HIVMDialect,
       mlir::triton::nvidia_gpu::TritonNvidiaGPUDialect,
       mlir::triton::gpu::TritonGPUDialect,
       mlir::triton::instrument::TritonInstrumentDialect,
