+++ python/triton/experimental/tle/language/builder.py
@@ -52,7 +52,7 @@ def setup_unified_builder_with_tle_builder(main_builder, buffer_builder):
         "create_dsa_extract_slice",
         "create_dsa_insert_slice",
         "create_dsa_subview",
-        # TileIR methods
+        # CommonIR methods
         'tile_get_buffer_type',
         'tile_get_tensor_type',
         'create_tile_alloc',
