+++ test/CommonIR/Ascend/success_case/matmul_double_buffer.py
@@ -23,7 +23,7 @@ from triton.experimental.tle.language.dsa import tile_alloc, tile_copy, tile_to_
 #    Even step: copy(next -> buf1), barrier, dot(buf0)
 #    Odd step:  copy(next -> buf0), barrier, dot(buf1)
 #
-#  NOTE: tileir_to_hivm merges tile.alloc of identical shape/dtype into a
+#  NOTE: commonir_to_hivm merges tile.alloc of identical shape/dtype into a
 #  single physical cbuf allocation. This causes buf0 and buf1 to share the
 #  same memref, breaking the double-buffer semantics. The copy overwrites the
 #  data that dot intends to read. This kernel produces WRONG results under
@@ -220,20 +220,21 @@ def dump_ttir(path=None, M=_DEFAULT_M, N=_DEFAULT_N, K=_DEFAULT_K, NUM_CORES=_DE
     return mlir
 
 
-def dump_tileir(path=None, M=_DEFAULT_M, N=_DEFAULT_N, K=_DEFAULT_K, NUM_CORES=_DEFAULT_NUM_CORES, return_module=False):
-    """Dump the frontend TileIR stage before TileIR-to-HIVM lowering."""
+def dump_commonir(path=None, M=_DEFAULT_M, N=_DEFAULT_N, K=_DEFAULT_K, NUM_CORES=_DEFAULT_NUM_CORES,
+                  return_module=False):
+    """Dump the frontend CommonIR stage before CommonIR-to-HIVM lowering."""
     if path is None:
-        path = os.path.join(os.path.dirname(os.path.abspath(__file__)), "matmul_double_buffer_tileir.mlir")
+        path = os.path.join(os.path.dirname(os.path.abspath(__file__)), "matmul_double_buffer_commonir.mlir")
 
     module = _compile_matmul_module(M=M, N=N, K=K, NUM_CORES=NUM_CORES)
     mlir = str(module)
     if "tile." not in mlir:
-        raise RuntimeError("dump_tileir: expected tile.* ops in frontend IR")
+        raise RuntimeError("dump_commonir: expected tile.* ops in frontend IR")
 
     if path:
         with open(path, "w") as f:
             f.write(mlir)
-        print(f"[dump_tileir] module.verify() = True; wrote TileIR ({len(mlir)} chars) to {path}")
+        print(f"[dump_commonir] module.verify() = True; wrote CommonIR ({len(mlir)} chars) to {path}")
 
     if return_module:
         return module
@@ -241,13 +242,13 @@ def dump_tileir(path=None, M=_DEFAULT_M, N=_DEFAULT_N, K=_DEFAULT_K, NUM_CORES=_
 
 
 # =============================================================================
-#  Full Linalg IR dump (TTIR -> TileIR -> Linalg lowering)
+#  Full Linalg IR dump (TTIR -> CommonIR -> Linalg lowering)
 # =============================================================================
 def dump_linalg(path=None, M=_DEFAULT_M, N=_DEFAULT_N, K=_DEFAULT_K, NUM_CORES=_DEFAULT_NUM_CORES):
-    """Compile matmul_kernel through full TileIR -> Linalg lowering pipeline.
+    """Compile matmul_kernel through full CommonIR -> Linalg lowering pipeline.
 
     Pipeline:
-      (1) tileir_to_hivm            -- tile.* -> memref/hivm
+      (1) commonir_to_hivm            -- tile.* -> memref/hivm
       (1b) erase_linalg_casts       -- eliminate unrealized casts
       (2) structure(r1) + discrete mask
       (3) unstructure + hivm + hfusion + llvm
@@ -268,12 +269,12 @@ def dump_linalg(path=None, M=_DEFAULT_M, N=_DEFAULT_N, K=_DEFAULT_K, NUM_CORES=_
     if path is None:
         path = os.path.join(os.path.dirname(os.path.abspath(__file__)), "matmul_double_buffer_linalg.mlir")
 
-    # -- (1) TileIR -> HIVM -----------------------------------------------
+    # -- (1) CommonIR -> HIVM -----------------------------------------------
     pm = ir.pass_manager(context)
     passes.common.add_inliner(pm)
-    ascend.passes.ttir.add_tileir_to_hivm(pm)
+    ascend.passes.ttir.add_commonir_to_hivm(pm)
     pm.run(module)
-    print(f"[dump_linalg] (1) tileir_to_hivm: verify={module.verify()}", flush=True)
+    print(f"[dump_linalg] (1) commonir_to_hivm: verify={module.verify()}", flush=True)
 
     # -- (1b) Erase unrealized_conversion_cast ops ------------------------
     # pm = ir.pass_manager(context)
@@ -364,8 +365,8 @@ if __name__ == "__main__":
     parser.add_argument("--no-check", action="store_true")
     parser.add_argument("--dump-ttir", nargs="?", const="", default=None,
                         help="Dump TTIR to PATH and exit; no device needed.")
-    parser.add_argument("--dump-tileir", nargs="?", const="", default=None,
-                        help="Dump frontend TileIR with tile.* ops to PATH and exit; no device needed.")
+    parser.add_argument("--dump-commonir", nargs="?", const="", default=None,
+                        help="Dump frontend CommonIR with tile.* ops to PATH and exit; no device needed.")
     parser.add_argument("--dump-linalg", nargs="?", const="", default=None,
                         help="Dump Linalg IR to PATH and exit; no device needed.")
     args = parser.parse_args()
@@ -377,8 +378,8 @@ if __name__ == "__main__":
         dump_ttir(path=(args.dump_ttir or None), M=M, N=N, K=K, NUM_CORES=num_cores)
         raise SystemExit(0)
 
-    if args.dump_tileir is not None:
-        dump_tileir(path=(args.dump_tileir or None), M=M, N=N, K=K, NUM_CORES=num_cores)
+    if args.dump_commonir is not None:
+        dump_commonir(path=(args.dump_commonir or None), M=M, N=N, K=K, NUM_CORES=num_cores)
         raise SystemExit(0)
 
     if args.dump_linalg is not None:
